+++ fugitive-darkness-provider/fugitive-darkness-provider-git/src/main/java/com/zer0s2m/fugitivedarkness/provider/git/impl/SearchInFileMatchCallable.java
@@ -32,6 +32,8 @@ class SearchInFileMatchCallable extends SearchInFileMatchFilterCallableAbstract<
         final AtomicInteger matcherCounterInFile = new AtomicInteger(0);
         final Matcher matcherCompile = pattern.matcher("");
 
+        long start = System.currentTimeMillis();
+
         try (final BufferedReader buf = new BufferedReader(reader)) {
             for (String line; (line = buf.readLine()) != null; ) {
                 lineNumber.set(lineNumber.get() + 1);
@@ -78,6 +80,13 @@ class SearchInFileMatchCallable extends SearchInFileMatchFilterCallableAbstract<
                 }
             }
         }
+
+        long finish = System.currentTimeMillis();
+
+        fileProcessingTime = finish - start;
+
+        StateEngineIOGitStatistics.TOTAL_PROCESSING_FILE.set(
+                StateEngineIOGitStatistics.TOTAL_PROCESSING_FILE.get() + fileProcessingTime);
 
         return collectPreviewCode(containerInfoSearchFileMatcherGitRepos, file);
     }