+++ fugitive-darkness-provider/fugitive-darkness-provider-git/src/main/java/com/zer0s2m/fugitivedarkness/provider/git/impl/StateEngineIOGitStatistics.java
@@ -1,0 +1,9 @@
+package com.zer0s2m.fugitivedarkness.provider.git.impl;
+
+import java.util.concurrent.atomic.AtomicLong;
+
+interface StateEngineIOGitStatistics {
+
+    AtomicLong TOTAL_PROCESSING_FILE = new AtomicLong(0);
+
+}